pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jasper"
version = "0.0.1"
description = "Financial research agent"
requires-python = ">=3.9"
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "langchain-core>=0.1.0",
    "langchain-openai>=0.1.0",
    "openai>=1.0.0",
    "httpx>=0.24.0",
    "python-dotenv>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["jasper*"]
